requirements.txt
aiohttp==3.7.3
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
bidict==0.21.2
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
google-api-core==1.26.0
google-auth==1.26.1
google-cloud-core==1.6.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
packaging==20.9
Pillow==8.1.0
proto-plus==1.13.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.6.0
pyparsing==2.4.7
python-engineio==4.0.0
python-socketio==5.0.4
pytz==2021.1
requests==2.25.1
rsa==4.7.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.3
watchdog==2.0.0
websocket-client==0.57.0
websockets==8.1
wrapt==1.12.1
yarl==1.6.3
